Restructured inference tasks to use mixins for parameter definitions.
